[APM] rename the index .apm-cm -> .apm-agent-configuration and many code identifiers to match the feature name (APM Agent Configuration)
