+++ x-pack/legacy/plugins/apm/server/lib/settings/agent_configuration/delete_configuration.ts
@@ -17,7 +17,7 @@ export async function deleteConfiguration({
 
   const params = {
     refresh: 'wait_for',
-    index: config.get<string>('apm_oss.cmIndex'),
+    index: config.get<string>('apm_oss.apmAgentConfigurationIndex'),
     id: configurationId
   };
