+++ x-pack/legacy/plugins/apm/server/lib/settings/agent_configuration/get_environments/index.ts
